:white_check_mark: Add test case
