+++ components/spin/__tests__/index.test.js
@@ -29,47 +29,6 @@ describe('Spin', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  describe('delay spinning', () => {
-    beforeAll(() => {
-      jest.useFakeTimers();
-    });
-
-    afterAll(() => {
-      jest.useRealTimers();
-    });
-
-    it("should render with delay when it's mounted with spinning=true and delay", () => {
-      const wrapper = mount(<Spin spinning delay={500} />);
-      expect(
-        wrapper
-          .find('.ant-spin')
-          .at(0)
-          .hasClass('ant-spin-spinning'),
-      ).toEqual(false);
-    });
-
-    it('should render when delay is init set', () => {
-      const wrapper = mount(<Spin spinning delay={100} />);
-
-      expect(
-        wrapper
-          .find('.ant-spin')
-          .at(0)
-          .hasClass('ant-spin-spinning'),
-      ).toEqual(false);
-
-      jest.runAllTimers();
-      wrapper.update();
-
-      expect(
-        wrapper
-          .find('.ant-spin')
-          .at(0)
-          .hasClass('ant-spin-spinning'),
-      ).toEqual(true);
-    });
-  });
-
   it('should be controlled by spinning', () => {
     const wrapper = mount(<Spin spinning={false} />);
     expect(wrapper.instance().state.spinning).toBe(false);
